Add files via upload
Python 3.9, pyQt5.14. PyQt5 application for read and write the coils and registers. This is the client (master) application ModBus on TCP. You may have to pip install pyModbusTCP if not done already. TCP IP address and port is hard coded. Only input Reg and coils are read and write from this application. UI is created on Designer and converted to PY file using pyuic5 -x main.ui -o main.py.